This pull request introduces a dependabot.yml configuration file to automate dependency updates within our repository. (#254)
* Created dependabot.yml

Enabling Dependabot version updates. This will run the Dependabot daily to check the versions if any new released & accordingly creates a PR with the latest version as well.

* Create dependabot.yml

This pull request adds a dependabot.yml configuration to automate daily dependency updates using GitHub's Dependabot. By maintaining a current and secure set of dependencies, we aim to enhance security and stability in our project.

* Update dependabot.yml

removed npm dependencies check as it was not required

---------

Co-authored-by: yashmeet29 <160218676+yashmeet29@users.noreply.github.com>