DINC0663527
This way  @ServiceName(value = "*", type = ApplicationService.class) of declaring things is the appropriate way. @ServiceName(*) doesn’t make any sense.
We use the same approach for all our built-in handlers in CAP Java itself, e.g. see all handlers here:
https://github.wdf.sap.corp/cds-java/cds-services/tree/main/cds-services-impl/src/main/java/com/sap/cds/services/impl/cds

This way just means that you register on all services that are served as APIs by the application.
There are many many other types of services in CAP (not only Messaging, but many more), which a plain @ServiceName(*) would all register to. This doesn’t make sense for an action handler.

This approach is also the right way completely independent if an application uses Messaging or not.